pythonProject/vcell-opt/pyproject.toml
[tool.poetry]
name = "vcell_opt"
version = "0.1.0"
description = ""
authors = ["Jim Schaff <schaff.jim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
copasi-basico = "0.40"
python-copasi = "4.37.264"
jupyter = "^1.0.0"
typer = {extras = ["all"], version = "^0.7.0"}
pytest = "^7.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
mypy = "^0.991"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
